+++ lib/firestore_model.dart
@@ -7,6 +7,10 @@ import 'package:flutter/widgets.dart';
 
 export 'src/firestore_model.dart';
 export 'src/settings.dart';
+export 'src/widgets/model_get_builder.dart';
+export 'src/widgets/model_single_builder.dart';
+export 'src/widgets/model_stream_get_builder.dart';
+export 'src/widgets/model_stream_single_builder.dart';
 
 class FirebaseApp {
   // Wait for Firebase to initialize