+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.5.0
+* modify widgets builders:
+  * `ModelSingleBuilder`: get `first` or `find` by docId.
+  * `ModelGetBuilder`: get documents with any `query`
+  * `ModelStreamGetBuilder`: stream `get` documents with any `query`
+  * `ModelStreamSingleBuilder`: stream `first` or `find` by docId.
+* fix read first document for `first` and `find`.
+
 ## 0.4.0
 * modify `find` && `first` to add document id to current model.
 * modify documents in code.